third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_word_03.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'word'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} word`
  );
}

function makeSample(offset){
  const kSample = '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right word
testIt('ltr', 0, 'right', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 1, 'right', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 2, 'right', '\n\u05D0^;\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 3, 'right', '\n\u05D0;^\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 4, 'right', '\n\u05D0;\u05D1^;\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 5, 'right', '\n\u05D0;\u05D1;^\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 6, 'right', '\n\u05D0;\u05D1;\u05D2^; \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 7, 'right', '\n\u05D0;\u05D1;\u05D2;^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 8, 'right', '\n\u05D0;\u05D1;\u05D2; ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 9, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 10, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 11, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 12, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 13, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 14, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 15, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 16, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\n');
testIt('ltr', 17, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\n');
testIt('ltr', 18, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\n');
testIt('ltr', 19, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\n');
testIt('ltr', 20, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|\n');
testIt('ltr', 21, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|\n');

// ltr left word
testIt('ltr', 0, 'left', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 1, 'left', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 2, 'left', '\n|\u05D0^;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 3, 'left', '\n|\u05D0;^\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 4, 'left', '\n\u05D0;|\u05D1^;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 5, 'left', '\n\u05D0;|\u05D1;^\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 6, 'left', '\n\u05D0;\u05D1;|\u05D2^; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 7, 'left', '\n\u05D0;\u05D1;|\u05D2;^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 8, 'left', '\n\u05D0;\u05D1;|\u05D2; ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 9, 'left', '\n\u05D0;\u05D1;\u05D2; |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 10, 'left', '\n\u05D0;\u05D1;\u05D2; |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 11, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 12, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 13, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 14, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 15, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 16, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\n');
testIt('ltr', 17, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\n');
testIt('ltr', 18, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\n');
testIt('ltr', 19, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\n');
testIt('ltr', 20, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\n');
testIt('ltr', 21, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\n');

// ltr forward word
testIt('ltr', 0, 'forward', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 1, 'forward', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 2, 'forward', '\n\u05D0^;\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 3, 'forward', '\n\u05D0;^\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 4, 'forward', '\n\u05D0;\u05D1^;\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 5, 'forward', '\n\u05D0;\u05D1;^\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 6, 'forward', '\n\u05D0;\u05D1;\u05D2^; \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 7, 'forward', '\n\u05D0;\u05D1;\u05D2;^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 8, 'forward', '\n\u05D0;\u05D1;\u05D2; ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 9, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 10, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 11, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 12, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 13, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 14, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 15, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\n');
testIt('ltr', 16, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\n');
testIt('ltr', 17, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\n');
testIt('ltr', 18, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\n');
testIt('ltr', 19, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\n');
testIt('ltr', 20, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|\n');
testIt('ltr', 21, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|\n');

// ltr backward word
testIt('ltr', 0, 'backward', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 1, 'backward', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 2, 'backward', '\n|\u05D0^;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 3, 'backward', '\n|\u05D0;^\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 4, 'backward', '\n\u05D0;|\u05D1^;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 5, 'backward', '\n\u05D0;|\u05D1;^\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 6, 'backward', '\n\u05D0;\u05D1;|\u05D2^; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 7, 'backward', '\n\u05D0;\u05D1;|\u05D2;^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 8, 'backward', '\n\u05D0;\u05D1;|\u05D2; ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 9, 'backward', '\n\u05D0;\u05D1;\u05D2; |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 10, 'backward', '\n\u05D0;\u05D1;\u05D2; |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 11, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 12, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 13, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 14, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 15, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\n');
testIt('ltr', 16, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\n');
testIt('ltr', 17, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\n');
testIt('ltr', 18, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\n');
testIt('ltr', 19, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\n');
testIt('ltr', 20, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\n');
testIt('ltr', 21, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\n');

// rtl right word
testIt('rtl', 0, 'right', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 1, 'right', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 2, 'right', '\n\u05D0^;\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 3, 'right', '\n\u05D0;^\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 4, 'right', '\n\u05D0;\u05D1^;\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 5, 'right', '\n\u05D0;\u05D1;^\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 6, 'right', '\n\u05D0;\u05D1;\u05D2^; \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 7, 'right', '\n\u05D0;\u05D1;\u05D2;^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 8, 'right', '\n\u05D0;\u05D1;\u05D2; ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 9, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 10, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 11, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 12, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 13, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 14, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 15, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 16, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\n');
testIt('rtl', 17, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\n');
testIt('rtl', 18, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\n');
testIt('rtl', 19, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\n');
testIt('rtl', 20, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|\n');
testIt('rtl', 21, 'right', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|\n');

// rtl left word
testIt('rtl', 0, 'left', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 1, 'left', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 2, 'left', '\n|\u05D0^;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 3, 'left', '\n|\u05D0;^\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 4, 'left', '\n\u05D0;|\u05D1^;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 5, 'left', '\n\u05D0;|\u05D1;^\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 6, 'left', '\n\u05D0;\u05D1;|\u05D2^; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 7, 'left', '\n\u05D0;\u05D1;|\u05D2;^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 8, 'left', '\n\u05D0;\u05D1;|\u05D2; ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 9, 'left', '\n\u05D0;\u05D1;\u05D2; |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 10, 'left', '\n\u05D0;\u05D1;\u05D2; |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 11, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 12, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 13, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 14, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 15, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 16, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\n');
testIt('rtl', 17, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\n');
testIt('rtl', 18, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\n');
testIt('rtl', 19, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\n');
testIt('rtl', 20, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\n');
testIt('rtl', 21, 'left', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\n');

// rtl forward word
testIt('rtl', 0, 'forward', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 1, 'forward', '\n^\u05D0|;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 2, 'forward', '\n\u05D0^;\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 3, 'forward', '\n\u05D0;^\u05D1|;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 4, 'forward', '\n\u05D0;\u05D1^;\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 5, 'forward', '\n\u05D0;\u05D1;^\u05D2|; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 6, 'forward', '\n\u05D0;\u05D1;\u05D2^; \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 7, 'forward', '\n\u05D0;\u05D1;\u05D2;^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 8, 'forward', '\n\u05D0;\u05D1;\u05D2; ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 9, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 10, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 11, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 12, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 13, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 14, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 15, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\n');
testIt('rtl', 16, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\n');
testIt('rtl', 17, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\n');
testIt('rtl', 18, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\n');
testIt('rtl', 19, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\n');
testIt('rtl', 20, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;|\n');
testIt('rtl', 21, 'forward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;|\n');

// rtl backward word
testIt('rtl', 0, 'backward', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 1, 'backward', '\n|\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 2, 'backward', '\n|\u05D0^;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 3, 'backward', '\n|\u05D0;^\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 4, 'backward', '\n\u05D0;|\u05D1^;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 5, 'backward', '\n\u05D0;|\u05D1;^\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 6, 'backward', '\n\u05D0;\u05D1;|\u05D2^; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 7, 'backward', '\n\u05D0;\u05D1;|\u05D2;^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 8, 'backward', '\n\u05D0;\u05D1;|\u05D2; ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 9, 'backward', '\n\u05D0;\u05D1;\u05D2; |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 10, 'backward', '\n\u05D0;\u05D1;\u05D2; |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 11, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 12, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 13, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 14, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 15, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\n');
testIt('rtl', 16, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\n');
testIt('rtl', 17, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\n');
testIt('rtl', 18, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\n');
testIt('rtl', 19, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\n');
testIt('rtl', 20, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\n');
testIt('rtl', 21, 'backward', '\n\u05D0;\u05D1;\u05D2; \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\n');
</script>
